Convert DeckGLMap component to typescript (#783)
* Convert DeckGlMap component to typescript
Closes #769

* correct proptypes

* Move setEditedData callback to userData prop

* Added COlorTables prop

* removed extra code
