+++ react/src/lib/components/DeckGLMap/components/DistanceScale.tsx
@@ -4,15 +4,15 @@ const convert = require("convert-units");
 
 interface scaleProps {
     // Needed the zoom value to calculate width in units
-    zoom: number;
+    zoom?: number;
     // Scale increment value
-    incrementValue: number;
+    incrementValue?: number | null;
     // Scale bar width in pixels per unit value
-    widthPerUnit: number;
+    widthPerUnit?: number | null;
     // positioning the scale ruler based on x and y values
-    position: number[];
+    position?: number[] | null;
     // default unit for the scale ruler
-    scaleUnit: string;
+    scaleUnit?: string;
 }
 
 const roundToStep = function (num: number, step: number) {
@@ -26,6 +26,8 @@ const DistanceScale: React.FC<scaleProps> = ({
     position,
     scaleUnit,
 }: scaleProps) => {
+    if (!zoom || !widthPerUnit || !incrementValue || !position) return null;
+
     const [rulerWidth, setRulerWidth] = React.useState<number>(0);
     const widthInUnits = widthPerUnit / Math.pow(2, zoom);
     const scaleRulerStyle = {
@@ -68,6 +70,7 @@ const DistanceScale: React.FC<scaleProps> = ({
 };
 
 DistanceScale.defaultProps = {
+    zoom: -3,
     incrementValue: 100,
     widthPerUnit: 100,
     position: [10, 10],
